fix server side handling
ComponentDidMount is not called on server side which is fine when adding
an window event listener. However ComponentDidMount was when the newly
added Device onChange was called. So moved the initial call of onChange
to the constructor so callers can get the event on the server side.

That also required separating the publish function from the onChange
handler. This way I can send just one listener to get the device info.